+++ apps/cowswap-frontend/src/modules/orders/hooks/useLoadMoreOrders.ts
@@ -1,10 +1,13 @@
-import { useAtom } from 'jotai'
-import { useCallback, useEffect, useMemo } from 'react'
+import { useAtom, useAtomValue } from 'jotai'
+import { useCallback, useMemo } from 'react'
 
-import { AMOUNT_OF_ORDERS_TO_FETCH } from '@cowprotocol/common-const'
+import { AMOUNT_OF_ORDERS_TO_FETCH, MAXIMUM_ORDERS_TO_FETCH } from '@cowprotocol/common-const'
 
-import { useApiOrders } from './useApiOrders'
+import { TabOrderTypes } from 'entities/routes/routes.atom'
 
+import { eoaTwapOrdersQueryAtom } from 'modules/twap/state/eoaTwapOrdersQueryAtom'
+
+import { apiOrdersAtom } from '../state/apiOrdersAtom'
 import { ordersLimitAtom } from '../state/ordersLimitAtom'
 
 interface UseLoadMoreOrdersReturn {
@@ -14,19 +17,23 @@ interface UseLoadMoreOrdersReturn {
   loadMore: () => void
 }
 
-export function useLoadMoreOrders(): UseLoadMoreOrdersReturn {
-  const [{ limit, isLoading }, setOrdersLimit] = useAtom(ordersLimitAtom)
-  const orders = useApiOrders()
-
-  useEffect(() => {
-    setOrdersLimit((prev) => ({ ...prev, isLoading: false }))
-  }, [orders, setOrdersLimit])
+export function useLoadMoreOrders(orderType: TabOrderTypes): UseLoadMoreOrdersReturn {
+  const isAdvancedOrders = orderType === TabOrderTypes.ADVANCED
+  const eoaTwapOrdersQuery = useAtomValue(eoaTwapOrdersQueryAtom)
+  const [limit, setLimit] = useAtom(ordersLimitAtom)
+  const { orders: apiOrders, isLoadingMore: apiOrdersLoadingMore } = useAtomValue(apiOrdersAtom)
+  const isLoading = isAdvancedOrders
+    ? eoaTwapOrdersQuery.isFetching && eoaTwapOrdersQuery.isPlaceholderData
+    : apiOrdersLoadingMore
 
   const loadMore = useCallback((): void => {
-    setOrdersLimit((prev) => ({ limit: prev.limit + AMOUNT_OF_ORDERS_TO_FETCH, isLoading: true }))
-  }, [setOrdersLimit])
+    setLimit((prev) =>
+      prev >= MAXIMUM_ORDERS_TO_FETCH ? prev : Math.min(prev + AMOUNT_OF_ORDERS_TO_FETCH, MAXIMUM_ORDERS_TO_FETCH),
+    )
+  }, [setLimit])
 
-  const hasMoreOrders = isLoading || orders.length >= limit
+  const hasMore = isAdvancedOrders ? limit < (eoaTwapOrdersQuery.data?.totalCount ?? 0) : apiOrders.length >= limit
+  const hasMoreOrders = isLoading || (limit < MAXIMUM_ORDERS_TO_FETCH && hasMore)
 
   return useMemo(
     () => ({