+++ libs/analytics/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [3.9.1](https://github.com/cowprotocol/cowswap/compare/analytics-v3.9.0...analytics-v3.9.1) (2026-08-12)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.13.0
+  - @cowprotocol/common-const@3.9.0
+  - @cowprotocol/common-hooks@3.9.1
+
+## [3.9.0](https://github.com/cowprotocol/cowswap/compare/analytics-v3.8.1...analytics-v3.9.0) (2026-08-05)
+
+### Minor Changes
+
+- feat(balances): track delegations for Solana (#7903)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-const@3.8.0
+  - @cowprotocol/common-hooks@3.9.0
+  - @cowprotocol/common-utils@3.12.0
+  - @cowprotocol/currency@1.3.0
+  - @cowprotocol/events@4.10.0
+  - @cowprotocol/types@4.8.0
+
 ## [3.8.1](https://github.com/cowprotocol/cowswap/compare/analytics-v3.8.0...analytics-v3.8.1) (2026-07-30)
 
 ### Patch Changes