+++ libs/common-hooks/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [3.9.1](https://github.com/cowprotocol/cowswap/compare/common-hooks-v3.9.0...common-hooks-v3.9.1) (2026-08-12)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.13.0
+  - @cowprotocol/common-const@3.9.0
+
+## [3.9.0](https://github.com/cowprotocol/cowswap/compare/common-hooks-v3.8.0...common-hooks-v3.9.0) (2026-08-05)
+
+### Minor Changes
+
+- feat(balances): track delegations for Solana (#7903)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-const@3.8.0
+  - @cowprotocol/common-utils@3.12.0
+  - @cowprotocol/types@4.8.0
+
 ## [3.8.0](https://github.com/cowprotocol/cowswap/compare/common-hooks-v3.7.0...common-hooks-v3.8.0) (2026-07-30)
 
 ### Minor Changes